refactor genericapiserver new to combine initialization
