+++ examples/apiserver/apiserver.go
@@ -66,7 +66,7 @@ func Run(serverOptions *genericoptions.ServerRunOptions) error {
 	genericvalidation.VerifyEtcdServersList(serverOptions)
 	config := genericapiserver.NewConfig(serverOptions)
 	config.Serializer = api.Codecs
-	s, err := genericapiserver.New(config)
+	s, err := config.New()
 	if err != nil {
 		return fmt.Errorf("Error in bringing up the server: %v", err)
 	}
