+++ federation/cmd/federation-apiserver/app/server.go
@@ -191,7 +191,7 @@ func Run(s *options.ServerRunOptions) error {
 		cachesize.SetWatchCacheSizes(s.WatchCacheSizes)
 	}
 
-	m, err := genericapiserver.New(genericConfig)
+	m, err := genericConfig.New()
 	if err != nil {
 		return err
 	}
